bench: share measurement profile status helper
